fixing issues

+++ Packager/Models/PodMetadataModels/Basics.cs
@@ -3,6 +3,6 @@ namespace Packager.Models.PodMetadataModels
     public class Basics
     {
         public string Format { get; set; }
-        public int Files { get; set; }
+        public long Files { get; set; }
     }
 }
